base: Extract base class for lineage base feature
So that it can be shared in multiple components
