+++ services/core/java/com/android/server/lineage/display/LiveDisplayFeature.java
@@ -15,10 +15,8 @@
  */
 package com.android.server.lineage.display;
 
-import android.content.ContentResolver;
 import android.content.Context;
 import android.hardware.display.ColorDisplayManager;
-import android.net.Uri;
 import android.os.Handler;
 import android.os.UserHandle;
 import android.util.Log;
@@ -26,8 +24,10 @@ import android.util.Log;
 import com.android.server.lineage.common.UserContentObserver;
 import com.android.server.lineage.display.LiveDisplayService.State;
 import com.android.server.lineage.display.TwilightTracker.TwilightState;
+import com.android.server.lineage.LineageBaseFeature;
+import com.android.server.lineage.display.LiveDisplayService.State;
+import com.android.server.lineage.display.TwilightTracker.TwilightState;
 
-import java.io.PrintWriter;
 import java.util.BitSet;
 
 import android.provider.Settings;
@@ -37,29 +37,19 @@ import static com.android.server.lineage.display.LiveDisplayService.DISPLAY_CHAN
 import static com.android.server.lineage.display.LiveDisplayService.MODE_CHANGED;
 import static com.android.server.lineage.display.LiveDisplayService.TWILIGHT_CHANGED;
 
-public abstract class LiveDisplayFeature {
+public abstract class LiveDisplayFeature extends LineageBaseFeature {
 
     protected static final String TAG = "LiveDisplay";
     protected static final boolean DEBUG = Log.isLoggable(TAG, Log.DEBUG);
 
-    protected final Context mContext;
-    protected final Handler mHandler;
     protected final boolean mNightDisplayAvailable;
 
-    private SettingsObserver mSettingsObserver;
     private State mState;
 
     public LiveDisplayFeature(Context context, Handler handler) {
-        mContext = context;
-        mHandler = handler;
+        super(context, handler);
         mNightDisplayAvailable = ColorDisplayManager.isNightDisplayAvailable(mContext);
     }
-
-    public abstract void onStart();
-
-    protected abstract void onSettingsChanged(Uri uri);
-
-    public abstract void dump(PrintWriter pw);
 
     public abstract boolean getCapabilities(final BitSet caps);
 
@@ -81,54 +71,10 @@ public abstract class LiveDisplayFeature {
         }
     }
 
-    void start() {
-        if (mSettingsObserver == null) {
-            mSettingsObserver = new SettingsObserver(mHandler);
-            onStart();
-        }
-    }
-
-    public void onDestroy() {
-        mSettingsObserver.unregister();
-    }
-
     protected void onScreenStateChanged() { }
 
     protected void onTwilightUpdated() { }
 
-    protected final void registerSettings(Uri... settings) {
-        mSettingsObserver.register(settings);
-    }
-
-    protected final boolean getBoolean(String setting, boolean defaultValue) {
-        return Settings.System.getIntForUser(mContext.getContentResolver(),
-                setting, (defaultValue ? 1 : 0), UserHandle.USER_CURRENT) == 1;
-    }
-
-    protected final void putBoolean(String setting, boolean value) {
-        Settings.System.putIntForUser(mContext.getContentResolver(),
-                setting, (value ? 1 : 0), UserHandle.USER_CURRENT);
-    }
-
-    protected final int getInt(String setting, int defaultValue) {
-        return Settings.System.getIntForUser(mContext.getContentResolver(),
-                setting, defaultValue, UserHandle.USER_CURRENT);
-    }
-
-    protected final void putInt(String setting, int value) {
-        Settings.System.putIntForUser(mContext.getContentResolver(),
-                setting, value, UserHandle.USER_CURRENT);
-    }
-
-    protected final String getString(String setting) {
-        return Settings.System.getStringForUser(mContext.getContentResolver(),
-                setting, UserHandle.USER_CURRENT);
-    }
-
-    protected final void putString(String setting, String value) {
-        Settings.System.putStringForUser(mContext.getContentResolver(),
-                setting, value, UserHandle.USER_CURRENT);
-    }
 
     protected final boolean isLowPowerMode() {
         return mState.mLowPowerMode;
@@ -149,36 +95,4 @@ public abstract class LiveDisplayFeature {
     public final boolean isNight() {
         return mState.mTwilight != null && mState.mTwilight.isNight();
     }
-
-    final class SettingsObserver extends UserContentObserver {
-
-        public SettingsObserver(Handler handler) {
-            super(handler);
-        }
-
-        public void register(Uri... uris) {
-            final ContentResolver cr = mContext.getContentResolver();
-            for (Uri uri : uris) {
-                cr.registerContentObserver(uri, false, this, UserHandle.USER_ALL);
-            }
-
-            observe();
-        }
-
-        public void unregister() {
-            mContext.getContentResolver().unregisterContentObserver(this);
-            unobserve();
-        }
-
-        @Override
-        protected void update() {
-            onSettingsChanged(null);
-        }
-
-        @Override
-        public void onChange(boolean selfChange, Uri uri) {
-            onSettingsChanged(uri);
-        }
-    }
-
 }
